working on refactoring for spawning requests
