+++ Cargo.toml
@@ -31,4 +31,3 @@ uuid = { version = "1.3.0", features = [
 ] }
 dotenvy = "0.15.6"
 lightning-invoice = "0.21.0"
-tonic_openssl_lnd = "0.2.0"
