+++ src/lightning/mod.rs
@@ -5,12 +5,6 @@ use dotenvy::var;
 use lightning_invoice::{Invoice, SignedRawInvoice};
 use std::str::FromStr;
 
-/// Decode a lightning invoice (bolt11)
-pub fn decode_invoice(payment_request: &str) -> Result<Invoice, MostroError> {
-    let invoice = Invoice::from_str(payment_request)?;
-
-    Ok(invoice)
-}
 
 /// Verify if an invoice is valid
 pub fn is_valid_invoice(payment_request: &str) -> Result<Invoice, MostroError> {
